up: makespace: Changes required for the Up!
This patch makes the changes to make the Up! Plus 3D printer work with a
Melzi board. The Melzi's bed FET is blown, so it is swapped for the fan
FET.

Data for the thermistors is included. This uses the original Up
thermistors with a copy of the original circuit calibrated to give
around 10 mV/deg. C